sync belts and sorters
code is redundant but I could not figure out how to get harmony to apply this method to both BuildTool_Inserter and BuildTool_Path.

Odd behavior to note here: BuildTool_Path seems to get patched regardless of the type of __instance. i.e if __instance is of type BuildTool_Path, BuildTool_Inserter, or BuildTool_Click, it will work. BuildTool_Inserter seems to only work when __instance is BuildTool_Inserter though, and this is the behavior I would expect.